python/langgraph-cloud-api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
